backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-admin-rangefilter
gunicorn
psycopg2-binary
dj-database-url